Fixed double-click error
A new error occurs when clicking two times or changing from one radio to
another without pressing the stop button. The stdout of the subprocess
is blocked and the process is killed. That used not to happen before.
Now I've removed the killing of the process when exiting from the while
loop in the stdout parsing function